requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
appdirs==1.4.3            # via zeep
arrow==0.13.1
attrs==19.1.0             # via zeep
cached-property==1.5.1    # via zeep
certifi==2019.3.9         # via requests
chardet==3.0.4            # via requests
click==7.0
defusedxml==0.6.0         # via zeep
idna==2.8                 # via requests
isodate==0.6.0            # via zeep
lxml==4.3.3               # via zeep
python-dateutil==2.8.0    # via arrow
python-dotenv==0.10.1
pytz==2019.1              # via zeep
requests-toolbelt==0.9.1  # via zeep
requests==2.21.0
six==1.12.0               # via isodate, python-dateutil, zeep
urllib3==1.24.2           # via requests
zeep==3.3.1
